Removing unused imports and casts
git-svn-id: https://svn.apache.org/repos/asf/cocoon/trunk@25814 13f79535-47bb-0310-9956-ffa450edef68

+++ src/blocks/midi/java/org/apache/cocoon/components/midi/xmidi/Utils.java
@@ -295,7 +295,7 @@ public final class Utils
     {
       int k = j - i - 1;
       int p = pow(128, i);
-      int m = (int) ba[k];
+      int m = ba[k];
       s += m * p;
       /*this.getLogger().debug(
       	"deltaToInt: in loop: s = "
